# src/docs/swagger.yaml
basePath: /
definitions:
...
  main.APICarRef:
    properties:
      car_id:
        description: CarID TeslaMate cars.id
        type: integer
      car_name:
        description: CarName 车辆昵称，可为空
        type: string
    type: object
...
  main.APIPositionRow:
    properties:
      battery_level:
        type: integer
      date:
        type: string
      driver_temp_setting:
        $ref: '#/definitions/main.NullFloat64'
      id:
        type: integer
      ideal_battery_range_km:
        type: number
      inside_temp:
        $ref: '#/definitions/main.NullFloat64'
      latitude:
        type: number
      longitude:
        type: number
      odometer:
        type: number
      outside_temp:
        $ref: '#/definitions/main.NullFloat64'
      power:
        type: integer
      rated_battery_range_km:
        type: number
      speed:
        type: integer
      usable_battery_level:
        $ref: '#/definitions/main.NullInt64'
    type: object
...
  main.APIStateRow:
    properties:
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
      state:
        type: string
    type: object
  main.APIUnitsLengthTemp:
    properties:
      unit_of_length:
        description: UnitsLength km 或 mi
        type: string
      unit_of_temperature:
        description: UnitsTemperature C 或 F
        type: string
    type: object
...
  main.APIUpdateRow:
    properties:
      end_date:
        type: string
      start_date:
        type: string
      update_id:
        type: integer
      version:
        type: string
    type: object
...
  main.NullFloat64:
    properties:
      float64:
        type: number
      valid:
        description: Valid is true if Float64 is not NULL
        type: boolean
    type: object
  main.NullInt64:
    properties:
      int64:
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
...
  main.RespPositionsData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      positions:
        items:
          $ref: '#/definitions/main.APIPositionRow'
        type: array
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
  main.RespPositionsList:
    properties:
      data:
        $ref: '#/definitions/main.RespPositionsData'
    type: object
...
  main.RespStatesData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      states:
        items:
          $ref: '#/definitions/main.APIStateRow'
        type: array
    type: object
  main.RespStatesList:
    properties:
      data:
        $ref: '#/definitions/main.RespStatesData'
    type: object
  main.RespUpdatesData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      updates:
        items:
          $ref: '#/definitions/main.APIUpdateRow'
        type: array
    type: object
  main.RespUpdatesList:
    properties:
      data:
        $ref: '#/definitions/main.RespUpdatesData'
    type: object
host: localhost:8080
info:
  contact: {}
  description: 与 TeslaMate 采集数据对齐的 JSON API；看板对照见仓库 `执行计划.md`。
  license:
    name: MIT
  title: TeslaMate API
  version: "1.0"
paths:
...
  /api/v1/cars/{CarID}/logging:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: PUT：TeslaMate 日志接口原始 JSON
          schema:
            items:
              type: integer
            type: array
      summary: TeslaMate 日志控制
      tags:
      - commands
  /api/v1/cars/{CarID}/logging/{Command}:
    put:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 子命令
        in: path
        name: Command
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PUT：TeslaMate 日志接口原始 JSON
          schema:
            items:
              type: integer
            type: array
      summary: TeslaMate 日志控制
      tags:
      - commands
...
  /api/v1/cars/{CarID}/positions:
    get:
      description: 查询 `positions` 表（时间序列与经纬度等）。支持 startDate、endDate 过滤 `date` 列，分页参数
        page、show（单页最大 5000）。
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 开始时间 RFC3339
        in: query
        name: startDate
        type: string
      - description: 结束时间 RFC3339
        in: query
        name: endDate
        type: string
      - default: 1
        description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - default: 500
        description: 每页条数
        in: query
        name: show
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespPositionsList'
      summary: 车辆位置与采样点
      tags:
      - dashboards
  /api/v1/cars/{CarID}/states:
    get:
      description: 查询 TeslaMate `states` 表（offline / asleep / online）。支持 startDate、endDate（RFC3339），默认按
        start_date 降序。
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 开始时间 RFC3339
        in: query
        name: startDate
        type: string
      - description: 结束时间 RFC3339
        in: query
        name: endDate
        type: string
      - default: 1
        description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - default: 100
        description: 每页条数
        in: query
        name: show
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespStatesList'
      summary: 车辆在线状态历史
      tags:
      - dashboards
...
  /api/v1/cars/{CarID}/updates:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespUpdatesList'
      summary: 软件更新记录
      tags:
      - cars
...
schemes:
- http
swagger: "2.0"